edited product renders after refresh
