+++ client/store/products.js
@@ -4,7 +4,6 @@ import axios from 'axios'
 const ALL_PRODUCTS = 'ALL_PRODUCTS'
 const ADD_PRODUCT = 'ADD_PRODUCT'
 const DELETE_PRODUCT = 'DELETE_PRODUCT'
-const EDIT_PRODUCT = 'EDIT_PRODUCT'
 const FILTER = 'FILTER'
 
 // Action Creators
@@ -20,11 +19,6 @@ export const deleteProduct = productId => ({
 
 export const addProduct = product => ({
   type: 'ADD_PRODUCT',
-  product
-})
-
-export const editProduct = product => ({
-  type: EDIT_PRODUCT,
   product
 })
 
@@ -59,19 +53,6 @@ export const removeProduct = id => {
   }
 }
 
-export const updateProduct = (id, product) => {
-  return async dispatch => {
-    try {
-      const {data} = await axios.patch(`/api/products/${id}`, product)
-      console.log(data)
-      dispatch(editProduct(data))
-      console.log(editProduct(data))
-    } catch (error) {
-      console.log(error)
-    }
-  }
-}
-
 // Add Product
 export const postProduct = product => {
   return async dispatch => {
@@ -94,8 +75,6 @@ const productsReducer = (state = initialState, action) => {
     case DELETE_PRODUCT:
       return state.filter(product => product.id !== action.productId)
     case ADD_PRODUCT:
-      return [...state, action.product]
-    case EDIT_PRODUCT:
       return [...state, action.product]
     case FILTER:
       return [...state].filter(product => product.category === action.category)
